tests: added test for add_id_request
